refactor: split per-language transforms into <Lang>IrTransformer.kt
Each <Lang>IrEmitter now imports its language-specific transformations
from a sibling <Lang>IrTransformer.kt file in the same package:

- JavaIrTransformer.kt
- KotlinIrTransformer.kt
- PythonIrTransformer.kt
- RustIrTransformer.kt
- ScalaIrTransformer.kt
- TypeScriptIrTransformer.kt
- SpringKotlinIrTransformer.kt

The emitter files now contain only the IrEmitter contract (sanitization
config, shared block, emit() overrides, sanitization string helpers,
companion keywords). All IR-tree transformations are top-level internal
extension functions in the transformer files.
