+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonIrEmitter.kt
@@ -185,60 +185,17 @@ open class PythonIrEmitter(
         union.convert()
             .sanitizeNames(sanitizationConfig)
 
-    override fun emit(refined: Refined): File {
-        val file = refined.convert()
-        val struct = file.findElement<Struct>()!!
-        val updatedStruct = struct.copy(
-            elements = struct.elements.mapNotNull { element ->
-                when {
-                    element is LanguageFunction && element.name == Name.of("validate") -> {
-                        val constraintExpr = refined.reference.convertConstraint(
-                            FieldCall(VariableReference(Name.of("self")), Name.of("value"))
-                        )
-                        function("validate") {
-                            arg("self", LanguageType.Custom(""))
-                            returnType(LanguageType.Boolean)
-                            returns(constraintExpr)
-                        }
-                    }
-                    element is LanguageFunction && element.name == Name.of("toString") -> {
-                        val toStringExpr = when (refined.reference.type) {
-                            is Reference.Primitive.Type.String -> "self.value"
-                            else -> "str(self.value)"
-                        }
-                        function("__str__") {
-                            arg("self", LanguageType.Custom(""))
-                            returnType(LanguageType.String)
-                            returns(RawExpression(toStringExpr))
-                        }
-                    }
-                    else -> element
-                }
-            },
-        )
-        return file
-            .transform {
-                matchingElements { _: Struct -> updatedStruct }
-            }
-            .sanitizeNames(sanitizationConfig)
-    }
+    override fun emit(refined: Refined): File = refined
+        .convert()
+        .replaceRefinedFunctions(refined)
+        .sanitizeNames(sanitizationConfig)
 
     override fun emit(endpoint: Endpoint): File {
         val endpointImports = endpoint.importReferences().distinctBy { it.value }
             .map { import("..model.${it.value}", it.value) }
-        val converted = endpoint.convert().findElement<Namespace>()!!
-        val flattened = converted.flattenNestedStructs()
-        val (moduleElements, classElements) = flattened.elements.partition { it is Struct || it is LanguageUnion }
-        val endpointClass = Namespace(
-            name = converted.name,
-            elements = classElements,
-            extends = converted.extends,
-        )
-        return LanguageFile(converted.name, buildList {
-            addAll(endpointImports)
-            addAll(moduleElements)
-            add(endpointClass)
-        })
+        return endpoint.convert()
+            .splitEndpointStructsToModuleLevel()
+            .let { it.copy(elements = endpointImports + it.elements) }
             .snakeCaseHandlerAndCallMethods()
             .sanitizeNames(sanitizationConfig)
     }
@@ -319,78 +276,6 @@ open class PythonIrEmitter(
         import(wirespecPath, "Wirespec"),
         import(wirespecPath, "_raise"),
     )
-
-    private fun <T : Element> T.snakeCaseHandlerAndCallMethods(): T = transform {
-        matchingElements { iface: Interface ->
-            if (iface.name == Name.of("Handler") || iface.name == Name.of("Call")) {
-                iface.copy(
-                    elements = iface.elements.map { element ->
-                        if (element is LanguageFunction) {
-                            element.copy(name = Name.of(element.name.snakeCase()))
-                        } else element
-                    },
-                )
-            } else iface
-        }
-    }
-
-    private fun <T : Element> T.flattenEndpointTypeRefs(endpointName: String): T = transform {
-        type { type, _ ->
-            if (type is LanguageType.Custom && type.name.startsWith("$endpointName.")) {
-                val suffix = type.name.removePrefix("$endpointName.")
-                if (suffix == "Call" || suffix == "Handler") type
-                else type.copy(name = suffix)
-            } else type
-        }
-    }
-
-    private fun <T : Element> T.addSelfReceiverToClientFields(): T {
-        val struct = (this as? File)?.findElement<Struct>()
-        val fieldNames = struct?.fields?.map { it.name.value() }?.toSet() ?: emptySet()
-        if (fieldNames.isEmpty()) return this
-
-        return transform {
-            statementAndExpression { stmt, tr ->
-                when (stmt) {
-                    is FieldCall -> {
-                        if (stmt.receiver == null && stmt.field.value() in fieldNames) {
-                            FieldCall(receiver = VariableReference(Name.of("self")), field = stmt.field)
-                        } else {
-                            FieldCall(
-                                receiver = stmt.receiver?.let { tr.transformExpression(it) },
-                                field = stmt.field,
-                            )
-                        }
-                    }
-                    else -> stmt.transformChildren(tr)
-                }
-            }
-        }
-    }
-
-    private fun <T : Element> T.snakeCaseClientFunctions(): T = transform {
-        matchingElements { func: LanguageFunction ->
-            func.copy(
-                name = Name.of(func.name.snakeCase()),
-                parameters = listOf(Parameter(Name.of("self"), LanguageType.Custom(""))) + func.parameters,
-            )
-        }
-        statementAndExpression { stmt, tr ->
-            when (stmt) {
-                is FunctionCall -> {
-                    val nameStr = stmt.name.value()
-                    val newName = if ("." in nameStr) stmt.name else Name.of(Name.of(nameStr).snakeCase())
-                    FunctionCall(
-                        name = newName,
-                        receiver = stmt.receiver?.let { tr.transformExpression(it) },
-                        arguments = stmt.arguments.mapValues { (_, v) -> tr.transformExpression(v) },
-                        isAwait = stmt.receiver != null,
-                    )
-                }
-                else -> stmt.transformChildren(tr)
-            }
-        }
-    }
 
     companion object : Keywords {
         override val reservedKeywords = setOf(
