+++ src/integration/spring/src/jvmMain/kotlin/community/flock/wirespec/integration/spring/kotlin/emit/SpringKotlinIrEmitter.kt
@@ -10,24 +10,13 @@ import community.flock.wirespec.compiler.core.parse.ast.Endpoint
 import community.flock.wirespec.compiler.core.parse.ast.Model
 import community.flock.wirespec.compiler.utils.Logger
 import community.flock.wirespec.emitters.kotlin.KotlinIrEmitter
-import community.flock.wirespec.ir.core.Interface
-import community.flock.wirespec.ir.core.Name
-import community.flock.wirespec.ir.core.RawElement
 import community.flock.wirespec.ir.core.file
-import community.flock.wirespec.ir.core.transform
 import community.flock.wirespec.ir.core.File as LanguageFile
 
 class SpringKotlinIrEmitter(packageName: PackageName) : KotlinIrEmitter(packageName, EmitShared(false)) {
 
-    override fun emit(endpoint: Endpoint): LanguageFile {
-        val annotation = endpoint.springAnnotation()
-        fun LanguageFile.injectSpringAnnotationBeforeHandler(): LanguageFile = transform {
-            injectBefore { iface: Interface ->
-                if (iface.name == Name.of("Handler")) listOf(RawElement(annotation)) else emptyList()
-            }
-        }
-        return super.emit(endpoint).injectSpringAnnotationBeforeHandler()
-    }
+    override fun emit(endpoint: Endpoint): LanguageFile = super.emit(endpoint)
+        .injectSpringAnnotationBeforeHandler(endpoint.springAnnotation())
 
     override fun emit(ast: AST, logger: Logger): NonEmptyList<Emitted> {
         val results = super.emit(ast, logger)
@@ -52,25 +41,6 @@ class SpringKotlinIrEmitter(packageName: PackageName) : KotlinIrEmitter(packageN
         val source = emitNativeHints(modelNames, endpointNames)
         val file = packageName.toDir() + "WirespecNativeHints." + extension.value
         return results + Emitted(file, source)
-    }
-
-    private fun Endpoint.springAnnotation(): String {
-        val pathString = "/" + path.joinToString("/") {
-            when (it) {
-                is Endpoint.Segment.Literal -> it.value
-                is Endpoint.Segment.Param -> "{${it.identifier.value}}"
-            }
-        }
-        return when (method) {
-            Endpoint.Method.GET -> "@org.springframework.web.bind.annotation.GetMapping(\"$pathString\")"
-            Endpoint.Method.POST -> "@org.springframework.web.bind.annotation.PostMapping(\"$pathString\")"
-            Endpoint.Method.PUT -> "@org.springframework.web.bind.annotation.PutMapping(\"$pathString\")"
-            Endpoint.Method.DELETE -> "@org.springframework.web.bind.annotation.DeleteMapping(\"$pathString\")"
-            Endpoint.Method.PATCH -> "@org.springframework.web.bind.annotation.PatchMapping(\"$pathString\")"
-            Endpoint.Method.OPTIONS -> "@org.springframework.web.bind.annotation.RequestMapping(value=[\"$pathString\"], method = [org.springframework.web.bind.annotation.RequestMethod.OPTIONS])"
-            Endpoint.Method.HEAD -> "@org.springframework.web.bind.annotation.RequestMapping(value=[\"$pathString\"], method = [org.springframework.web.bind.annotation.RequestMethod.HEAD])"
-            Endpoint.Method.TRACE -> "@org.springframework.web.bind.annotation.RequestMapping(value=[\"$pathString\"], method = [org.springframework.web.bind.annotation.RequestMethod.TRACE])"
-        }
     }
 
     private fun emitNativeHints(modelNames: List<String>, endpointNames: List<String>): String {
